requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==24.2
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
